packages/llm-provider-py/pyproject.toml
[project]
name = "deployai-llm-provider-py"
version = "0.0.0"
description = "Epic 5: Anthropic + OpenAI + failover LLM providers (Python)."
requires-python = ">=3.13,<3.14"
dependencies = [
  "httpx>=0.28.0",
  "opentelemetry-api>=1.28.0",
  "pyyaml>=6.0.0",
]

[project.optional-dependencies]
aws = ["boto3>=1.35.0"]

[dependency-groups]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.7.0",
  "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_provider_py"]

[tool.hatch.build.targets.sdist]
include = ["src/llm_provider_py", "pyproject.toml", "scripts", "tests"]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N", "RUF", "ASYNC"]

[tool.mypy]
python_version = "3.13"
strict = true
mypy_path = "src"
packages = ["llm_provider_py"]

[[tool.mypy.overrides]]
module = "boto3"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
